Test converging on shared vulnerable dependency
`documentation@4.0.0` and `gulp-sass@3.1.0` have been added as
dependencies of the project. Between them they share a dependency on
`hoek@2.16.3`, which has known vulnerabilities. That makes this a good
test for how a system handles dependencies that appear in multiple
places in the dependency tree. `nsp` reports the same vulnerability
each time the vulnerable module appears in the dependency tree, but also
reports the dependency path used to reach the vulnerable module.